Implemented concurrency-safe transfer engine and async fraud detection

+++ src/main/java/com/finflow/repository/TransactionRepository.java
@@ -60,5 +60,5 @@ public interface TransactionRepository extends JpaRepository<Transaction, Long> 
             WHERE t.fromWallet.id = :walletId
               AND t.status = com.finflow.enums.TransactionStatus.COMPLETED
             """)
-    Double avgTransactionAmount(@Param("walletId") Long walletId);
+    BigDecimal avgTransactionAmount(@Param("walletId") Long walletId);
 }
